Decouple the RPC handling from the replication server/client -> move to ImStorage for now PART 3
