+++ src/rpc/server.cpp
@@ -1,4 +1,4 @@
-// Copyright 2022 Memgraph Ltd.
+// Copyright 2023 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -21,6 +21,8 @@ bool Server::Start() { return server_.Start(); }
 void Server::Shutdown() { server_.Shutdown(); }
 
 void Server::AwaitShutdown() { server_.AwaitShutdown(); }
+
+bool Server::IsRunning() { return server_.IsRunning(); }
 
 const io::network::Endpoint &Server::endpoint() const { return server_.endpoint(); }
 }  // namespace memgraph::rpc
