+++ src/storage/v2/CMakeLists.txt
@@ -37,7 +37,9 @@ set(storage_v2_src_files
     replication/slk.cpp
     replication/replication_persistence_helper.cpp
     replication/rpc.cpp
-    replication/replication.cpp)
+    replication/replication.cpp
+    inmemory/replication/replication_server.cpp
+)
 
 #######################
 find_package(gflags REQUIRED)
